requirements.txt
langchain
langchain-openai
langchain-anthropic
langchain-ollama
pydantic
python-dotenv
requests
python-whois
rich
pyfiglet
dnspython
fpdf2
questionary
chromadb
langchain-chroma
langchain-text-splitters